+++ packages/node/src/app.module.ts
@@ -5,6 +5,7 @@ import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ScheduleModule } from '@nestjs/schedule';
 import { DbModule } from '@subql/node-core';
+import { AdminModule } from './admin/admin.module';
 import { ConfigureModule } from './configure/configure.module';
 import { FetchModule } from './indexer/fetch.module';
 import { MetaModule } from './meta/meta.module';
@@ -17,6 +18,7 @@ import { MetaModule } from './meta/meta.module';
     ScheduleModule.forRoot(),
     FetchModule,
     MetaModule,
+    AdminModule,
   ],
   controllers: [],
 })