+++ packages/node/src/indexer/fetch.module.ts
@@ -17,6 +17,7 @@ import {
   PoiSyncService,
   InMemoryCacheService,
   SandboxService,
+  MonitorService,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { EthereumApiConnection } from '../ethereum/api.connection';
@@ -82,6 +83,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
         dynamicDsService: DynamicDsService,
         unfinalizedBlocks: UnfinalizedBlocksService,
         connectionPoolState: ConnectionPoolStateManager<EthereumApiConnection>,
+        monitorService?: MonitorService,
       ) =>
         nodeConfig.workers
           ? new WorkerBlockDispatcherService(
@@ -97,6 +99,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
               dynamicDsService,
               unfinalizedBlocks,
               connectionPoolState,
+              monitorService,
             )
           : new BlockDispatcherService(
               apiService,
@@ -125,6 +128,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
         DynamicDsService,
         UnfinalizedBlocksService,
         ConnectionPoolStateManager,
+        MonitorService,
       ],
     },
     FetchService,
@@ -141,8 +145,9 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
       useClass: ProjectService,
       provide: 'IProjectService',
     },
+    MonitorService,
     UnfinalizedBlocksService,
   ],
-  exports: [StoreService, StoreCacheService],
+  exports: [StoreService, StoreCacheService, MonitorService, PoiService],
 })
 export class FetchModule {}